Add aconfig_declarations block to SDK target.
Without this block we're not correctly generating the set of flags for
revert-annotations-exportable.txt, meaning APIs that should be in the
finalized JAR are missing. This shouldn't be necessary after
aosp/3001319, but something seems not to be working as expected. In
order to unblock API finalization, just add the declarations block for
now.

Bug: 335454564
Test: Look at revert-annotations-exportable.txt, see all relevant flags
Change-Id: I0ae35108bca7b368af2457068b1a7e6dbf597dcf